ExerciseList and Controller Updates
Completely redesigned how the ExerciseList class saves and loads the exercises.

Made a new menu for viewing the exercises.

Added four new exercises: Abs, Biceps, Cardio, and Triceps

Added a way for the exercise lists to be  sorted when items are added to it.

Added helper methods to a lot of methods and cut down a ton of duplicate code (adding/removing exercises, etc.)