JavaFX Stats and Apache POI Stats
Started wokring on interacting with Microsoft Excel spreadsheets with my program. Will most likely be used to write information and not read it.

Changed Exercise class to include max weight, reps, volume, and one rep max estimate. Will use these on the graphs in the future.

Changed reps instance variable in Set class from int to double. Needed to do this in order to perform the division required for one rep max estimation