JFileChooser and Excel Graphs
Created a JFileChooser object to save the Excel graphs. Need to implement a loop to ensure that the user can select and deselect a file as many times as they want.

Created two more Excel graphs for statistical data: Max Reps and Max Volume. Need to attempt to make a loop to create all 3 graphs to cut down on duplicate code.