Saving WorkoutTracker in Excel
Started coming up with a way to save the current WorkoutTracker object in Excel. Each workout will be bordered off and exercises in them will be color coded.

Currently, only prints out in a long line but will eventually be across the columns as well.

Created a FileChooser for the user to select a directory for the WorkoutTracker folder to be created with files automatically created and saved in it. Trying to figure out the best way to save information on the Excel sheets without having too many of them, but also not having too many sheets or the sheets being cluttered.